Declare hard dependency on doctrine/cache
This package has a hard dependency on doctrine/cache v1 via `CacheFactory`.
But that dependency was not declared as such, yet it was fullfilled by coincidence
via a transient dependency. And the relatively recent doctrine/cache v2
introduced **a lot** of breaking changes that break not only this package
tests, but also virtually all of dependents of this package.

Fixes #46